@@ -4,25 +4,25 @@ publish_to: "none"
 version: 1.0.0+1
 
 environment:
-  sdk: ^3.1.0
+  sdk: ^3.7.0
 
 dependencies:
+  ansi_escape_codes: ^2.1.0
+  bloc: ^9.1.0
+  cupertino_icons: ^1.0.8
+  equatable: ^2.0.7
   flutter:
     sdk: flutter
-  scopo:
-    path: ../..
-  cupertino_icons: ^1.0.8
-  bloc: ^9.1.0
-  equatable: ^2.0.7
-  ansi_escape_codes: ^2.1.0
-  shared_preferences: ^2.5.3
   flutter_markdown_plus: ^1.0.5
   flutter_syntax_view: ^4.1.7
+  scopo:
+    path: ../..
+  shared_preferences: ^2.5.3
 
 dev_dependencies:
+  flutter_lints: ^5.0.0
   flutter_test:
     sdk: flutter
-  flutter_lints: ^5.0.0
 
 flutter:
   uses-material-design: true
