@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:scopo_demo/common/presentation/blinking_box.dart';
+
+import 'blinking_box.dart';
 
 class Box extends StatelessWidget {
-  final Widget title;
-  final Color? titleBackgroundColor;
-  final Color? titleForegroundColor;
+  final Color? borderColor;
   final Color? backgroundColor;
   final Color? foregroundColor;
   final Color? blinkingColor;
@@ -13,9 +12,7 @@ class Box extends StatelessWidget {
 
   const Box({
     super.key,
-    required this.title,
-    this.titleBackgroundColor,
-    this.titleForegroundColor,
+    this.borderColor,
     this.backgroundColor,
     this.foregroundColor,
     this.blinkingColor,
@@ -25,59 +22,28 @@ class Box extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final titleBackgroundColor =
-        this.titleBackgroundColor ?? Theme.of(context).colorScheme.primary;
-    Widget content = Padding(
-      padding: contentPadding,
-      child: child,
-    );
+    Widget content = Padding(padding: contentPadding, child: child);
     if (blinkingColor case final blinkingColor?) {
-      content = BlinkingBox(
-        blinkingColor: blinkingColor,
-        child: content,
-      );
+      content = BlinkingBox(blinkingColor: blinkingColor, child: content);
     }
 
     return Container(
       decoration: BoxDecoration(
         color: backgroundColor ?? Theme.of(context).colorScheme.surface,
         borderRadius: BorderRadius.circular(8),
-        border: Border.all(color: titleBackgroundColor),
+        border: Border.all(
+          color:
+              borderColor ??
+              backgroundColor ??
+              Theme.of(context).colorScheme.primary,
+        ),
       ),
       clipBehavior: Clip.antiAlias,
-      child: Table(
-        defaultColumnWidth: IntrinsicColumnWidth(),
-        children: [
-          TableRow(
-            children: [
-              ColoredBox(
-                color: titleBackgroundColor,
-                child: Padding(
-                  padding: const EdgeInsets.symmetric(
-                    vertical: 0,
-                    horizontal: 8,
-                  ),
-                  child: DefaultTextStyle(
-                    style: DefaultTextStyle.of(context).style.copyWith(
-                          color: titleForegroundColor ??
-                              Theme.of(context).colorScheme.onPrimary,
-                        ),
-                    child: Center(child: title),
-                  ),
-                ),
-              ),
-            ],
-          ),
-          TableRow(children: [
-            DefaultTextStyle(
-              style: DefaultTextStyle.of(context).style.copyWith(
-                    color: foregroundColor ??
-                        Theme.of(context).colorScheme.onSurface,
-                  ),
-              child: content,
-            )
-          ]),
-        ],
+      child: DefaultTextStyle(
+        style: DefaultTextStyle.of(context).style.copyWith(
+          color: foregroundColor ?? Theme.of(context).colorScheme.onSurface,
+        ),
+        child: content,
       ),
     );
   }
