@@ -1,19 +1,21 @@
 import 'package:flutter/material.dart';
 
-import 'box.dart';
+import 'titled_box.dart';
 
 class Consumer extends StatefulWidget {
   final Widget title;
-  final List<Widget> description;
-  final String Function(BuildContext context) consumerValue;
-  final Color blinkingColor;
+  final Widget? description;
+  final List<Widget>? descriptions;
+  final Widget Function(BuildContext context) builder;
+  final Color? blinkingColor;
 
   const Consumer({
     super.key,
     required this.title,
-    required this.description,
-    required this.consumerValue,
-    required this.blinkingColor,
+    this.description,
+    this.descriptions,
+    required this.builder,
+    this.blinkingColor,
   });
 
   @override
@@ -22,31 +24,31 @@ class Consumer extends StatefulWidget {
 
 class _ConsumerState extends State<Consumer> {
   @override
-  Widget build(BuildContext context) => Box(
-        title: widget.title,
-        blinkingColor: widget.blinkingColor,
-        child: DefaultTextStyle(
-          style: Theme.of(context).textTheme.bodySmall!,
-          child: IntrinsicHeight(
-            child: Row(
-              children: [
-                Expanded(
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.stretch,
-                    children: widget.description,
+  Widget build(BuildContext context) => TitledBox(
+    title: widget.title,
+    blinkingColor: widget.blinkingColor,
+    child: DefaultTextStyle(
+      style: Theme.of(context).textTheme.bodySmall!,
+      child: IntrinsicHeight(
+        child: Row(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Expanded(
+              child:
+                  widget.description ??
+                  Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: widget.descriptions ?? [],
                   ),
-                ),
-                VerticalDivider(
-                  width: 20,
-                  color: Theme.of(context).colorScheme.primary,
-                ),
-                Text(
-                  widget.consumerValue(context),
-                  style: Theme.of(context).textTheme.titleSmall,
-                ),
-              ],
             ),
-          ),
+            const VerticalDivider(),
+            DefaultTextStyle(
+              style: Theme.of(context).textTheme.titleSmall!,
+              child: widget.builder(context),
+            ),
+          ],
         ),
-      );
+      ),
+    ),
+  );
 }
