@@ -1,14 +1,14 @@
 import 'dart:math';
 
 import 'package:scopo/scopo.dart';
-import 'package:scopo_demo/common/data/real_services/key_value_service.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../common/domain/fake_exception.dart';
 import '../common/data/fake_services/analytics.dart';
 import '../common/data/fake_services/connectivity.dart';
 import '../common/data/fake_services/http_client.dart';
 import '../common/data/fake_services/some_controller.dart';
+import '../common/data/real_services/key_value_service.dart';
+import '../common/domain/fake_exception.dart';
 import '../utils/app_environment.dart';
 import 'app.dart';
 
@@ -30,10 +30,8 @@ class AppDeps implements ScopeDeps {
     required this.someController,
   }) : _sharedPreferences = sharedPreferences;
 
-  KeyValueService keyValueService(String prefix) => KeyValueService(
-        sharedPreferences: _sharedPreferences,
-        prefix: prefix,
-      );
+  KeyValueService keyValueService(String prefix) =>
+      KeyValueService(sharedPreferences: _sharedPreferences, prefix: prefix);
 
   /// Method uses [DoubleProgressIterator] to track and report granular
   /// initialization progress ([double] 0.0 to 1.0).
@@ -71,7 +69,7 @@ class AppDeps implements ScopeDeps {
       randomFakeError('$SharedPreferences initialization error');
       yield ScopeProgress(progressIterator.nextProgress());
 
-      httpClient = HttpClient();
+      httpClient = const HttpClient();
       await httpClient.init();
       randomFakeError('$HttpClient initialization error');
       yield ScopeProgress(progressIterator.nextProgress());
@@ -81,7 +79,7 @@ class AppDeps implements ScopeDeps {
       randomFakeError('$Connectivity initialization error');
       yield ScopeProgress(progressIterator.nextProgress());
 
-      analytics = Analytics();
+      analytics = const Analytics();
       await analytics.init();
       randomFakeError('$Analytics initialization error');
       yield ScopeProgress(progressIterator.nextProgress());
