@@ -9,25 +9,23 @@ class SplashScreen extends StatelessWidget {
   const SplashScreen({super.key, required this.progress});
 
   @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        title: Text('$SplashScreen'),
-        backgroundColor: Colors.lightBlue,
-      ),
-      backgroundColor: Colors.lightBlueAccent,
-      body: Padding(
-        padding: EdgeInsets.all(8),
-        child: Center(
-          child: SizedBox(
-            width: 100,
-            child: AnimatedProgressIndicator(
-              value: progress,
-              builder: (value) => LinearProgressIndicator(value: value),
-            ),
+  Widget build(BuildContext context) => Scaffold(
+    appBar: AppBar(
+      title: Text('$SplashScreen'),
+      backgroundColor: Colors.lightBlue,
+    ),
+    backgroundColor: Colors.lightBlueAccent,
+    body: Padding(
+      padding: const EdgeInsets.all(8),
+      child: Center(
+        child: SizedBox(
+          width: 100,
+          child: AnimatedProgressIndicator(
+            value: progress,
+            builder: (value) => LinearProgressIndicator(value: value),
           ),
         ),
       ),
-    );
-  }
+    ),
+  );
 }
