@@ -1,9 +1,11 @@
+import '../../../utils/app_environment.dart';
+
 class SomeController {
   Future<void> init() async {
-    await Future<void>.delayed(const Duration(milliseconds: 500));
+    await Future<void>.delayed(AppEnvironment.defaultPause);
   }
 
   Future<void> dispose() async {
-    await Future<void>.delayed(const Duration(milliseconds: 1000));
+    await Future<void>.delayed(AppEnvironment.defaultPause);
   }
 }
