@@ -1,4 +1,4 @@
-final class FakeException {
+final class FakeException implements Exception {
   final String message;
 
   FakeException(this.message);
