@@ -13,24 +13,18 @@ class Counter extends StatelessWidget {
   });
 
   @override
-  Widget build(BuildContext context) {
-    return DefaultTextStyle(
-      style: Theme.of(context).textTheme.titleLarge!,
-      child: Row(
-        children: [
-          Text('$title: '),
-          Expanded(
-              child: Text(
-            '$value',
-            textAlign: TextAlign.end,
-          )),
-          IconButton(
-            color: Theme.of(context).colorScheme.primary,
-            onPressed: increment,
-            icon: const Icon(Icons.add_circle),
-          ),
-        ],
-      ),
-    );
-  }
+  Widget build(BuildContext context) => DefaultTextStyle(
+    style: Theme.of(context).textTheme.titleLarge!,
+    child: Row(
+      children: [
+        Text('$title: '),
+        Expanded(child: Text('$value', textAlign: TextAlign.end)),
+        IconButton(
+          color: Theme.of(context).colorScheme.primary,
+          onPressed: increment,
+          icon: const Icon(Icons.add_circle),
+        ),
+      ],
+    ),
+  );
 }
