@@ -12,19 +12,13 @@ class MarkdownBlock extends StatelessWidget {
     this.data, {
     super.key,
     this.fontSize,
-    this.selectable = false,
+    this.selectable = true,
     this.indent = 8,
   });
 
   @override
-  Widget build(BuildContext context) {
-    return Padding(
-      padding: EdgeInsets.only(top: indent),
-      child: Markdown(
-        data,
-        selectable: selectable,
-        fontSize: fontSize,
-      ),
-    );
-  }
+  Widget build(BuildContext context) => Padding(
+    padding: EdgeInsets.only(top: indent),
+    child: Markdown(data, selectable: selectable, fontSize: fontSize),
+  );
 }
