@@ -38,14 +38,10 @@ class SizedTabBar extends TabBar {
     super.textScaler,
     super.indicatorAnimation,
   }) : super(
-            tabs: tabs
-                .map(
-                  (e) => SizedBox(
-                    height: height,
-                    child: e,
-                  ),
-                )
-                .toList(growable: false));
+         tabs: tabs
+             .map((e) => SizedBox(height: height, child: e))
+             .toList(growable: false),
+       );
 
   @override
   Size get preferredSize => Size.fromHeight(height + indicatorWeight);
