@@ -3,13 +3,13 @@ import 'package:scopo/scopo.dart';
 
 import '../app/app.dart';
 import '../app/app_error.dart';
-import '../app/theme_manager.dart';
-import '../common/presentation/animated_progress_indicator.dart';
+import '../app/theme_manager/theme_manager.dart';
 import '../common/data/fake_services/some_bloc.dart';
 import '../common/data/fake_services/some_controller.dart';
+import '../common/presentation/animated_progress_indicator.dart';
 import '../common/presentation/sized_tab_bar.dart';
-import '../scope_provider_demo.dart/scope_provider_demo.dart';
-import '../scope_state_manager_demo/scope_state_manager_demo.dart';
+import '../data_access_demo/data_access_demo.dart';
+import '../scope_initializer_demo/scope_state_manager_demo.dart';
 import 'home_counter.dart';
 import 'home_deps.dart';
 import 'home_navigation_block.dart';
@@ -28,7 +28,7 @@ final class Home extends Scope<Home, HomeDeps, HomeContent> {
     super.tag,
     required ScopeInitFunction<double, HomeDeps> super.init,
     this.isRoot = true,
-  });
+  }) : super(pauseAfterInitialization: const Duration(milliseconds: 500));
 
   /// Provides access the scope params, i.e. to the widget [Home].
   static Home paramsOf(BuildContext context, {bool listen = true}) =>
@@ -84,12 +84,14 @@ class HomeAppBar extends AppBar {
                 ThemeManager.of(context, listen: false).toggleBrightness();
               },
               onLongPress: () {
-                ThemeManager.of(context, listen: false).mode = ThemeMode.system;
+                ThemeManager.of(context, listen: false).resetBrightness();
               },
-              icon: Icon(switch (ThemeManager.of(context).brightness) {
-                Brightness.dark => Icons.light_mode,
-                Brightness.light => Icons.dark_mode,
-              }),
+              icon: Icon(
+                switch (ThemeManager.select(context, (m) => m.brightness)) {
+                  Brightness.dark => Icons.light_mode,
+                  Brightness.light => Icons.dark_mode,
+                },
+              ),
             ),
           ],
           bottom: withTabs
@@ -102,9 +104,9 @@ class HomeAppBar extends AppBar {
                       .colorScheme
                       .onPrimary
                       .withValues(alpha: 0.5),
-                  tabs: [
-                    Tab(text: 'ScopeProvider'),
-                    Tab(text: 'ScopeStateManager'),
+                  tabs: const [
+                    Tab(text: 'Data access'),
+                    Tab(text: 'Async initialization'),
                     Tab(text: 'Other'),
                   ],
                 )
@@ -125,20 +127,18 @@ class _FakeContent extends StatefulWidget {
 
 class _FakeContentState extends State<_FakeContent> {
   @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: HomeAppBar(context, withTabs: false),
-      body: Padding(
-        padding: EdgeInsets.all(8),
-        child: Center(
-          child: AnimatedProgressIndicator(
-            value: widget.progress,
-            builder: (value) => CircularProgressIndicator(value: value),
+  Widget build(BuildContext context) => Scaffold(
+        appBar: HomeAppBar(context, withTabs: false),
+        body: Padding(
+          padding: const EdgeInsets.all(8),
+          child: Center(
+            child: AnimatedProgressIndicator(
+              value: widget.progress,
+              builder: (value) => CircularProgressIndicator(value: value),
+            ),
           ),
         ),
-      ),
-    );
-  }
+      );
 }
 
 mixin ChangeNotifierOnState<T extends StatefulWidget> on State<T>
@@ -184,13 +184,11 @@ final class HomeContent extends ScopeContent<Home, HomeDeps, HomeContent> {
     notifyListeners();
   }
 
-  void openDialog(BuildContext context) {
-    showAdaptiveDialog(
-      useRootNavigator: false,
-      context: context,
-      barrierDismissible: true,
-      builder: (context) {
-        return Dialog(
+  Future<void> openDialog(BuildContext context) => showAdaptiveDialog<void>(
+        useRootNavigator: false,
+        context: context,
+        barrierDismissible: true,
+        builder: (context) => Dialog(
           backgroundColor: Theme.of(
             context,
           ).colorScheme.surface.withValues(alpha: 0.7),
@@ -199,86 +197,80 @@ final class HomeContent extends ScopeContent<Home, HomeDeps, HomeContent> {
             child: Column(
               mainAxisSize: MainAxisSize.min,
               children: [
-                AppBar(title: Text('Dialog'), primary: false),
-                SizedBox(height: 20),
-                HomeCounter(),
-                SizedBox(height: 20),
+                AppBar(
+                  title: const Text('Dialog'),
+                  primary: false,
+                ),
+                const SizedBox(height: 20),
+                const HomeCounter(),
+                const SizedBox(height: 20),
               ],
             ),
           ),
-        );
-      },
-    );
-  }
+        ),
+      );
 
   void openBottomSheet(BuildContext context) {
     showBottomSheet(
       context: context,
       clipBehavior: Clip.antiAlias,
-      builder: (context) {
-        return SafeArea(
-          child: Column(
-            mainAxisSize: MainAxisSize.min,
-            children: [
-              AppBar(title: Text('Bottom sheet')),
-              SizedBox(height: 20),
-              HomeCounter(),
-              SizedBox(height: 20),
-              ElevatedButton(
-                onPressed: () {
-                  openBottomSheet(context);
-                },
-                child: Text('more'),
-              ),
-              SizedBox(height: 20),
-            ],
-          ),
-        );
-      },
+      builder: (context) => SafeArea(
+        child: Column(
+          mainAxisSize: MainAxisSize.min,
+          children: [
+            AppBar(title: const Text('Bottom sheet')),
+            const SizedBox(height: 20),
+            const HomeCounter(),
+            const SizedBox(height: 20),
+            ElevatedButton(
+              onPressed: () {
+                openBottomSheet(context);
+              },
+              child: const Text('more'),
+            ),
+            const SizedBox(height: 20),
+          ],
+        ),
+      ),
     );
   }
 
-  void openModalBottomSheet(BuildContext context) {
-    showModalBottomSheet(
-      context: context,
-      clipBehavior: Clip.antiAlias,
-      builder: (context) {
-        return SafeArea(
+  Future<void> openModalBottomSheet(BuildContext context) =>
+      showModalBottomSheet(
+        context: context,
+        clipBehavior: Clip.antiAlias,
+        builder: (context) => SafeArea(
           child: Column(
             mainAxisSize: MainAxisSize.min,
             children: [
-              AppBar(title: Text('Bottom sheet')),
-              SizedBox(height: 20),
-              HomeCounter(),
-              SizedBox(height: 20),
+              AppBar(title: const Text('Bottom sheet')),
+              const SizedBox(height: 20),
+              const HomeCounter(),
+              const SizedBox(height: 20),
             ],
           ),
-        );
-      },
-    );
-  }
+        ),
+      );
 
   @override
-  Widget build(BuildContext context) {
-    return Builder(
-      builder: (context) {
-        return DefaultTabController(
+  Widget build(BuildContext context) => Builder(
+        builder: (context) => DefaultTabController(
           length: 3,
           child: Scaffold(
             appBar: HomeAppBar(context),
             body: TabBarView(
               children: [
-                ScopeProviderDemo(),
-                ScopeStateManagerDemo(),
+                const DataAccessDemo(),
+                const ScopeStateBuilderDemo(),
                 ListView(
                   children: [
-                    Text('You have pushed this buttons many times:'),
-                    HomeCounter(),
-                    SizedBox(height: 20),
-                    SizedBox(height: 20),
-                    SizedBox(height: 20),
-                    HomeNavigationBlock(),
-                    SizedBox(height: 20),
+                    const Text('You have pushed this buttons many times:'),
+                    const HomeCounter(),
+                    const SizedBox(height: 20),
+                    const SizedBox(height: 20),
+                    const SizedBox(height: 20),
+                    const HomeNavigationBlock(),
+                    const SizedBox(height: 20),
                     ElevatedButton(
                       onPressed: () {
                         setState(() {});
@@ -290,8 +282,6 @@ final class HomeContent extends ScopeContent<Home, HomeDeps, HomeContent> {
               ],
             ),
           ),
-        );
-      },
-    );
-  }
+        ),
+      );
 }
