@@ -1,13 +1,13 @@
 export 'src/environment/scope_config.dart'
-    hide d, e, RawScopeLogCallback, source;
+    hide RawScopeLogCallback, d, e, source;
 export 'src/listenable/listenable_selector.dart';
-export 'src/scope_fork/scope_fork.dart';
 export 'src/navigation_node/navigation_node.dart';
 export 'src/progress_iterator/double_progress_iterator.dart';
-export 'src/scope_provider/scope_provider.dart';
-export 'src/scope_state_manager/scope_state_manager.dart';
+export 'src/scope/scope.dart';
 export 'src/scope/scope_consumer.dart';
 export 'src/scope/scope_deps.dart';
-export 'src/scope/scope.dart';
+export 'src/scope_fork/scope_fork.dart';
+export 'src/scope_initializer/scope_initializer.dart';
+export 'src/scope_provider/scope_provider.dart';
 export 'src/utils/compare_utils.dart';
 export 'src/utils/notifier.dart';
