@@ -5,7 +5,7 @@ import 'listen.dart';
 
 /// A widget that rebuilds when a value selected from a `Listenable` changes.
 ///
-/// See [ListenableSelectExtension.listenTo].
+/// See [ListenableSelectExtension.select].
 class ListenableSelector<L extends Listenable, T extends Object?>
     extends StatefulWidget {
   final L listenable;
