@@ -1,97 +1,85 @@
 part of '../scope_provider.dart';
 
 abstract base class ScopeNotifierBottom<
-    W extends ScopeNotifierBottom<W, E, T>,
-    E extends ScopeNotifierElementBase<W, E, T>,
-    T extends ChangeNotifier> extends ScopeListenableProviderBottom<W, E, T> {
+    W extends ScopeNotifierBottom<W, E, M>,
+    E extends ScopeNotifierElementBase<W, E, M>,
+    M extends Listenable> extends ScopeProviderBottom<W, E, M> {
   const ScopeNotifierBottom({
     super.key,
-    required super.create,
-    super.debugString,
-    required super.builder,
   });
 
-  const ScopeNotifierBottom.value({
-    super.key,
-    required super.value,
-    super.debugString,
-    required super.builder,
-  }) : super.value();
-
-  const ScopeNotifierBottom.raw({
-    super.key,
-    super.debugString,
-    required super.builder,
-  }) : super.raw();
+  // +++
+  // @override
+  // bool updateShouldNotify(ScopeNotifierBottom<W, E, T> oldWidget) => false;
 
   @override
-  E createElement();
+  E createScopeElement();
 
-  static E? maybeGet<
-          W extends ScopeNotifierBottom<W, E, T>,
-          E extends ScopeNotifierElementBase<W, E, T>,
-          T extends ChangeNotifier>(
+  static C? maybeOf<W extends ScopeInheritedWidget,
+          C extends ScopeContext<W, M>, M extends Listenable>(
+    BuildContext context, {
+    required bool listen,
+  }) =>
+      ScopeProviderBottom.maybeOf<W, C, M>(context, listen: listen);
+
+  static C of<W extends ScopeInheritedWidget, C extends ScopeContext<W, M>,
+          M extends Listenable>(
+    BuildContext context, {
+    required bool listen,
+  }) =>
+      ScopeProviderBottom.of<W, C, M>(context, listen: listen);
+
+  static V select<W extends ScopeInheritedWidget, C extends ScopeContext<W, M>,
+          M extends Listenable, V extends Object?>(
     BuildContext context,
+    V Function(C context) selector,
   ) =>
-      ScopeListenableProviderBottom.maybeGet<W, E, T>(context);
-
-  static E get<
-          W extends ScopeNotifierBottom<W, E, T>,
-          E extends ScopeNotifierElementBase<W, E, T>,
-          T extends ChangeNotifier>(
-    BuildContext context,
-  ) =>
-      ScopeListenableProviderBottom.get<W, E, T>(context);
-
-  static E? maybeDepend<
-          W extends ScopeNotifierBottom<W, E, T>,
-          E extends ScopeNotifierElementBase<W, E, T>,
-          T extends ChangeNotifier>(
-    BuildContext context,
-  ) =>
-      ScopeProviderBottom.maybeDepend<W, E, T>(context);
-
-  static E depend<
-          W extends ScopeNotifierBottom<W, E, T>,
-          E extends ScopeNotifierElementBase<W, E, T>,
-          T extends ChangeNotifier>(
-    BuildContext context,
-  ) =>
-      ScopeProviderBottom.depend<W, E, T>(context);
-
-  static V? maybeSelect<
-          W extends ScopeNotifierBottom<W, E, T>,
-          E extends ScopeNotifierElementBase<W, E, T>,
-          T extends ChangeNotifier,
-          V extends Object?>(
-    BuildContext context,
-    V Function(E) selector,
-  ) =>
-      ScopeListenableProviderBottom.maybeSelect<W, E, T, V>(context, selector);
-
-  static V select<
-          W extends ScopeNotifierBottom<W, E, T>,
-          E extends ScopeNotifierElementBase<W, E, T>,
-          T extends ChangeNotifier,
-          V extends Object?>(
-    BuildContext context,
-    V Function(E) selector,
-  ) =>
-      ScopeListenableProviderBottom.select<W, E, T, V>(context, selector);
-
-  @override
-  String toStringShort() =>
-      debugString?.call() ?? '${ScopeNotifierBottom<W, E, T>}';
+      ScopeProviderBottom.select<W, C, M, V>(context, selector);
 }
 
-base class ScopeNotifierElementBase<W extends ScopeNotifierBottom<W, E, T>,
-        E extends ScopeNotifierElementBase<W, E, T>, T extends ChangeNotifier>
-    extends ScopeListenableProviderElementBase<W, E, T> {
+abstract base class ScopeNotifierElementBase<
+        W extends ScopeProviderBottom<W, E, M>,
+        E extends ScopeNotifierElementBase<W, E, M>,
+        M extends Listenable> extends ScopeProviderElementBase<W, E, M>
+    implements ScopeContext<W, M> {
   ScopeNotifierElementBase(super.widget);
+  var _shouldNotify = false;
+  var _updateChild = true;
 
   @override
-  void unmount() {
-    super.unmount();
-    _model?.dispose();
+  void init() {
+    model.addListener(_listener);
+    super.init();
   }
+
+  @override
+  void dispose() {
+    super.dispose();
+    model.removeListener(_listener);
+  }
+
+  void _listener() {
+    _shouldNotify = true;
+    markNeedsBuild();
+  }
+
+  @override
+  void notifyClients(W oldWidget) {
+    super.notifyClients(oldWidget);
+    _shouldNotify = false;
+  }
+
+  @override
+  void performRebuild() {
+    if (_shouldNotify) {
+      notifyClients(widget);
+      _updateChild = dependencies?.contains(this) ?? false;
+    }
+    super.performRebuild();
+    _updateChild = true;
+  }
+
+  @override
+  Element? updateChild(Element? child, Widget? newWidget, Object? newSlot) =>
+      _updateChild ? super.updateChild(child, newWidget, newSlot) : child;
 }
