@@ -1,13 +1,11 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 
-part 'scope_listenable_provider/scope_listenable_provider.dart';
-part 'scope_listenable_provider/scope_listenable_provider_base.dart';
-part 'scope_listenable_provider/scope_listenable_provider_bottom.dart';
-part 'scope_notifier/scope_notifier.dart';
-part 'scope_notifier/scope_notifier_base.dart';
-part 'scope_notifier/scope_notifier_bottom.dart';
+part 'scope_provider/base.dart';
+part 'scope_provider/interface.dart';
 part 'scope_provider/scope_provider.dart';
 part 'scope_provider/scope_provider_base.dart';
 part 'scope_provider/scope_provider_bottom.dart';
-part 'scope_provider/scope_provider_facade.dart';
+part 'scope_notifier/scope_notifier.dart';
+part 'scope_notifier/scope_notifier_base.dart';
+part 'scope_notifier/scope_notifier_bottom.dart';
