@@ -1,56 +1,56 @@
 part of '../scope_provider.dart';
 
-final class ScopeProvider<T extends Object>
-    extends ScopeProviderBase<ScopeProvider<T>, T> {
+final class ScopeProvider<M extends Object>
+    extends ScopeProviderBase<ScopeProvider<M>, M>
+    with _ScopeInheritedWidgetMixin<M> {
+  @override
+  final Widget Function(BuildContext context) builder;
+
+  @override
+  final String? debugName;
+
   const ScopeProvider({
     super.key,
     required super.create,
-    super.dispose,
-    super.debugString,
-    required super.builder,
+    required super.dispose,
+    required this.builder,
+    this.debugName,
   });
 
   const ScopeProvider.value({
     super.key,
     required super.value,
-    super.debugString,
-    required super.builder,
+    required this.builder,
+    this.debugName,
   }) : super.value();
 
-  static T? maybeGet<T extends Object>(
-    BuildContext context,
-  ) =>
-      ScopeProviderBase.maybeGet<ScopeProvider<T>, T>(context)?.model;
+  static M? maybeOf<M extends Object>(
+    BuildContext context, {
+    required bool listen,
+  }) =>
+      ScopeProviderBottom.maybeOf<ScopeProvider<M>,
+          ScopeContext<ScopeProvider<M>, M>, M>(
+        context,
+        listen: listen,
+      )?.model;
 
-  static T get<T extends Object>(
-    BuildContext context,
-  ) =>
-      ScopeProviderBase.get<ScopeProvider<T>, T>(context).model;
+  static M of<M extends Object>(
+    BuildContext context, {
+    required bool listen,
+  }) =>
+      ScopeProviderBottom.of<ScopeProvider<M>,
+          ScopeContext<ScopeProvider<M>, M>, M>(
+        context,
+        listen: listen,
+      ).model;
 
-  static T? maybeDepend<T extends Object>(
+  static V select<M extends Object, V extends Object?>(
     BuildContext context,
+    V Function(M model) selector,
   ) =>
-      ScopeProviderBase.maybeDepend<ScopeProvider<T>, T>(context)?.model;
-
-  static T depend<T extends Object>(
-    BuildContext context,
-  ) =>
-      ScopeProviderBase.depend<ScopeProvider<T>, T>(context).model;
-
-  static V? maybeSelect<T extends Object, V extends Object?>(
-    BuildContext context,
-    V Function(T) selector,
-  ) =>
-      ScopeProviderBase.maybeSelect<ScopeProvider<T>, T, V>(
-          context, (f) => selector(f.model));
-
-  static V select<T extends Object, V extends Object?>(
-    BuildContext context,
-    V Function(T) selector,
-  ) =>
-      ScopeProviderBase.select<ScopeProvider<T>, T, V>(
-          context, (f) => selector(f.model));
-
-  @override
-  String toStringShort() => debugString?.call() ?? '${ScopeProvider<T>}';
+      ScopeProviderBottom.select<ScopeProvider<M>,
+          ScopeContext<ScopeProvider<M>, M>, M, V>(
+        context,
+        (context) => selector(context.model),
+      );
 }
